+++ cmd/houston/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/ternaryss/houston/internal/app/db"
+	"github.com/ternaryss/houston/internal/app/handlers"
 	"github.com/ternaryss/houston/internal/app/settings"
 	"github.com/ternaryss/houston/internal/app/web"
 )
@@ -10,6 +11,10 @@ func main() {
 	settings := settings.LoadSettings()
 	dbProvider := db.NewDbProvider(settings)
 	defer dbProvider.CloseConnection()
-	server := web.NewServer(settings)
+	usersStore := db.NewUsersStore(dbProvider.Db())
+	errorsHandler := handlers.NewErrorsHandler()
+	dashboardHandler := handlers.NewDashboardHandler()
+	usersHandler := handlers.NewUsersHandler(usersStore)
+	server := web.NewServer(settings, errorsHandler, dashboardHandler, usersHandler)
 	server.Run()
 }