+++ internal/app/handlers/dashboard.go
@@ -1,0 +1,19 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/ternaryss/houston/internal/app/helpers"
+	"github.com/ternaryss/houston/views"
+)
+
+type DashboardHandler struct{}
+
+func NewDashboardHandler() *DashboardHandler {
+	return &DashboardHandler{}
+}
+
+func (h *DashboardHandler) Dashboard(res http.ResponseWriter, req *http.Request) {
+	template := views.Dashboard()
+	helpers.RenderPage(template, res, req)
+}