+++ CHANGELOG.md
@@ -4,5 +4,7 @@ History of **Houston we have (no) problem** application with changes description
 
 ## Next release
 
+- `FEATURE` Sign up
+- `FEATURE` Request execution time middleware
 - `FEATURE` Containerization with Docker
 - `FEATURE` Initialize application (GoLang, SQLite, Templ, TailwindCSS, FlyonUI, HTMX, Air)